pyproject.toml
[project]
name = "polymarket-scanner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.3",
    "numpy>=2.4.3",
    "requests>=2.32.5",
    "python-dotenv>=1.0.1",
    "py-clob-client",
]
